+++ Changelog.rst
@@ -1,9 +1,10 @@
-
-version NEXTRELEASE
+version NEXTVERSION
 -------------------
 
 **2024-??-??**
 
+* Include the UM version as a field property when reading UM files
+  (https://github.com/NCAS-CMS/cf-python/issues/777)
 * Fix bug where `cf.example_fields` returned a `list`
   of Fields rather than a `Fieldlist`
   (https://github.com/NCAS-CMS/cf-python/issues/725)